bottlescan/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
pillow>=10.0.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
